feat(desktop): Rename file to startPageManagement
