+++ desktop-app/app/constants/startPageManagement.js
